Update indeterministic test in orm_to_json_utils_tests
The original code here does a next(iter(<dictionary>)) which has
an indeterministic order, so this occasionally returns the
variantTag for an AIP variant, which gets transformed with a
different key, causing this test to intermittently fail. For this
reason, we'll hard select the the variant 
VT1726961_2103343353_r0390_100 to look at the keys for.